+++ packages/tts-ggml/test/mobile/integration.auto.cjs
@@ -66,6 +66,10 @@ async function runParlerTest (options = {}) { // eslint-disable-line no-unused-v
   return runIntegrationModule('../integration/parler.test.js', options)
 }
 
+async function runPocketIntegrationTest (options = {}) { // eslint-disable-line no-unused-vars
+  return runIntegrationModule('../integration/pocket.integration.test.js', options)
+}
+
 async function runRtfBenchmarkTest (options = {}) { // eslint-disable-line no-unused-vars
   return runIntegrationModule('../integration/rtf-benchmark.test.js', options)
 }
@@ -106,6 +110,7 @@ module.exports = {
   runOutputSampleRateTest,
   runParlerWerTest,
   runParlerTest,
+  runPocketIntegrationTest,
   runRtfBenchmarkTest,
   runStreamingBenchmarkTest,
   runSupertonicMtlTest,